requirements.txt
Pillow
pywin32
opencv-python
screeninfo
